Add from_extractor and deprecate extractor_middleware
